# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cafe Backend API",
        "description": "Dokumentasi API Cafe Backend",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Admin Orders"
                ],
                "summary": "Ambil semua pesanan",
                "description": "Mengambil seluruh data pesanan beserta transaksi, diurutkan dari yang terbaru",
                "operationId": "bd28d4677d3217255664864122c057f3",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil daftar pesanan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Admin Orders"
                ],
                "summary": "Ambil detail pesanan",
                "description": "Mengambil detail pesanan berdasarkan ID",
                "operationId": "c1f954b8c64b3af1c15d9ea61232cf73",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Order",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pesanan ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin Orders"
                ],
                "summary": "Hapus pesanan",
                "description": "Menghapus pesanan yang belum completed",
                "operationId": "2ad07571ebb91cc422288404d5f523bf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pesanan berhasil dihapus"
                    },
                    "403": {
                        "description": "Pesanan yang sudah selesai tidak boleh dihapus"
                    },
                    "404": {
                        "description": "Order tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{id}/status": {
            "patch": {
                "tags": [
                    "Admin Orders"
                ],
                "summary": "Update status pesanan",
                "description": "Mengubah status order (pending, processing, ready, completed, cancelled)",
                "operationId": "de13e19ed91dc2454329b8ee3193ab70",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "processing",
                                            "ready",
                                            "completed",
                                            "cancelled"
                                        ],
                                        "example": "processing"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status berhasil diperbarui"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Order tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{orderId}/item/{index}": {
            "patch": {
                "tags": [
                    "Admin Orders"
                ],
                "summary": "Update status item dalam order",
                "description": "Mengubah status satu item di dalam field JSON items",
                "operationId": "5c212fc775825aea1232bbafffb7deab",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "path",
                        "description": "ID Order",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "index",
                        "in": "path",
                        "description": "Index item di array items",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "description": "Status item (default: completed)",
                                        "type": "string",
                                        "example": "completed"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status item berhasil diperbarui"
                    },
                    "404": {
                        "description": "Order atau item tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token JWT dengan format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}